Fix optional bootstrap steps

+++ bootstrap.sh
@@ -10,7 +10,7 @@ echo "= Dotfiles Setup ="
 if echo $platform | grep -q Darwin; then
   echo "== macOS Installation =="
 
-  if ! command brew >/dev/null 2>&1; then
+  if ! command -v brew >/dev/null 2>&1; then
     echo "=== Installing Homebrew ==="
     ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
   fi
@@ -18,7 +18,7 @@ if echo $platform | grep -q Darwin; then
   brew tap Homebrew/bundle
   brew bundle
 
-  sudo mv /etc/zshenv /etc/zshrc
+  [[ -f /etc/zshenv ]] && sudo mv /etc/zshenv /etc/zshrc
   # TODO: move /usr/local/bin to the top of /etc/paths
 elif echo $platform | grep -q Ubuntu; then
   echo "== Ubuntu Installation =="
